feat: 농작업 삭제 요청 시 status가 `COMPLETE` 이면 실패하도록 변경

+++ backend/src/main/java/com/imsnacks/Nyeoreumnagi/work/service/MyWorkService.java
@@ -64,6 +64,11 @@ public class MyWorkService {
         if (myWork.getMember().getId() != memberId) {
             throw new WorkException(MY_WORK_NOT_FOUND);
         }
+
+        if(!myWork.isDone()){
+            throw new WorkException(MY_WORK_NOT_COMPLETED);
+        }
+
         myWorkRepository.deleteById(request.myWorkId());
     }
